Android and VOTD
I finally sorted out all of the little bugs to deal with working on
Android.  It turns out that you have to be really careful with how you
open/close windows on Android.  It’s tricky.  I also added a verse of
the day, pulling in a simple RSS parser.